8620 modulenotfounderror no module named onnxscript in test py3x 311 pipeline (#8638)
Fixes #8620  .

### Description

Adds `onnxscript` as an explicit dependency.

I have tried to find where this onnxscript package was coming from
before. For that, I tried all Python versions from 3.9 to 3.12, all
versions of onnxruntime and onnx_graphsurgeon, and all versions later
than 1.13.0 of onnx.

None of these would include `onnxscript`.

I suppose that this was a requirement of another library and was removed
in some new version.
I don't think it is worth the trouble of further investigating to find
which package it was, since we wouldn't want to freeze a package version
for this reason. So, instead, I propose we just add onnxscript as a
dependency.

### Potential issue

I am not sure if this will trigger the running of the ONNX tests in
Python < 3.10 and how it will impact those tests.

A few sentences describing the changes proposed in this pull request.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

---------

Signed-off-by: R. Garcia-Dias <rafaelagd@gmail.com>
Signed-off-by: Rafael Garcia-Dias <rafaelagd@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>